# pkg/router/handlers/pet.go
/*
 * Swagger Petstore - OpenAPI 3.0
 *
 * This is a sample Pet Store Server based on the OpenAPI 3.0 specification.  You can find out more about Swagger at [https://swagger.io](https://swagger.io). In the third iteration of the pet store, we've switched to the design first approach! You can now help us improve the API whether it's by making changes to the definition itself or to the code. That way, with time, we can improve the API in general, and expose some of the new features in OAS3.  _If you're looking for the Swagger 2.0/OAS 2.0 version of Petstore, then click [here](https://editor.swagger.io/?url=https://petstore.swagger.io/v2/swagger.yaml). Alternatively, you can load via the `Edit > Load Petstore OAS 2.0` menu option!_  Some useful links: - [The Pet Store repository](https://github.com/swagger-api/swagger-petstore) - [The source API definition for the Pet Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)
 *
 * API version: 1.0.11
 * Contact: apiteam@swagger.io
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package handlers

import (
	"encoding/json"
	"errors"
	"fmt"
	"io"
	"log"
	"net/http"
	"strconv"

	"github.com/gorilla/mux"
	"github.com/renan-campos/petstore/pkg/router/handlers/models"
)

// Service Code Start
// For this exercise, I'm ignoring data persistence.
type petID int64

var localID petID

func GeneratePetID() petID {
	out := localID
	localID += 1
	return out
}

func strToPetID(strID string) (petID, error) {
	intID, err := strconv.Atoi(strID)
	if err != nil {
		return 0, err
	}
	return petID(intID), nil
}

var PetNotFound = errors.New("pet not found")
var pets map[petID]models.Pet = make(map[petID]models.Pet)

// Service Code End

func AddPet(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json; charset=UTF-8")

	rawPet, err := io.ReadAll(r.Body)
	if err != nil {
		w.WriteHeader(http.StatusInternalServerError)
		log.Printf("error reading json: %v\n", err)
		return
	}

	var pet models.Pet
	if err := json.Unmarshal(rawPet, &pet); err != nil {
		w.WriteHeader(http.StatusInternalServerError)
		log.Printf("error unmarshalling json: %v\n", err)
		return
	}

	serviceCode := func() error {
		id := GeneratePetID()
		pet.Id = int64(id)
		pets[id] = pet
		return nil
	}
	if err := serviceCode(); err != nil {
		w.WriteHeader(http.StatusInternalServerError)
		log.Printf("error in service: %v\n", err)
		return
	}

	outPet, err := json.Marshal(pet)
	if err != nil {
		w.WriteHeader(http.StatusInternalServerError)
		log.Printf("error marshalling pet: %v\n", err)
		return
	}

	if _, err := w.Write(outPet); err != nil {
		w.WriteHeader(http.StatusInternalServerError)
		log.Printf("error writing response body: %v\n", err)
		return
	}
}

func GetPetById(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json; charset=UTF-8")

	urlVars := mux.Vars(r)
	id, ok := urlVars["petId"]
	if !ok {
		w.WriteHeader(http.StatusInternalServerError)
		log.Printf("Could not find id in url path\n")
		return
	}

	// Service Code Start
	serviceGetById := func(id string) (models.Pet, error) {
		var pet models.Pet
		petID, err := strToPetID(id)
		if err != nil {
			return pet, fmt.Errorf("failed to convert pet id: %v", err)
		}
		pet, ok := pets[petID]
		if !ok {
			return pet, PetNotFound
		}
		return pet, nil
	}
	// Service Code End
	pet, err := serviceGetById(id)

	if err != nil {
		switch {
		case errors.Is(err, PetNotFound):
			w.WriteHeader(http.StatusNotFound)
		default:
			w.WriteHeader(http.StatusInternalServerError)
		}
		return
	}

	outPet, err := json.Marshal(pet)
	if err != nil {
		w.WriteHeader(http.StatusInternalServerError)
		log.Printf("failed to marshal pet: %v\n", err)
		return
	}
	if _, err := w.Write(outPet); err != nil {
		w.WriteHeader(http.StatusInternalServerError)
		log.Printf("failed to write pet: %v\n", err)
		return
	}
	w.WriteHeader(http.StatusOK)
	return
}

func DeletePet(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json; charset=UTF-8")

	urlVars := mux.Vars(r)
	id, ok := urlVars["petId"]
	if !ok {
		w.WriteHeader(http.StatusInternalServerError)
		log.Printf("could not get petId parameter from call\n")
		return
	}

	// Service Code Start
	serviceDeleteById := func(id string) error {
		petID, err := strToPetID(id)
		if err != nil {
			return fmt.Errorf("failed to convert id: %v")
		}
		if _, ok := pets[petID]; !ok {
			return PetNotFound
		}
		delete(pets, petID)
		return nil
	}

	err := serviceDeleteById(id)
	if err != nil {
		switch {
		case errors.Is(err, PetNotFound):
			w.WriteHeader(http.StatusNotFound)
		default:
			w.WriteHeader(http.StatusInternalServerError)
			log.Printf("failed to delete pet: %v\n", err)
		}
		return
	}
	w.WriteHeader(http.StatusNoContent)
	return
}

func FindPetsByStatus(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
	w.WriteHeader(http.StatusNotImplemented)
}

func FindPetsByTags(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
	w.WriteHeader(http.StatusNotImplemented)
}

func UpdatePet(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
	w.WriteHeader(http.StatusNotImplemented)
}

func UpdatePetWithForm(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
	w.WriteHeader(http.StatusNotImplemented)
}

func UploadFile(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
	w.WriteHeader(http.StatusNotImplemented)
}